+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `QUILL_RDTSC_CLOCK` option is replaced by `QUILL_CHRONO_CLOCK` which is by OFF by default.
 * Improve compiler error message when trying to log a non copy constructible user defined type
 * Fix buffer reallocation bug on TimestampFormatter. In previous versions any timestamp format set to 'set_pattern' expanding to a string longer than 32 bytes would cause a crash. ([#24](https://github.com/odygrd/quill/issues/24))
+* Added a latency benchmark, easily extendable for any logger
 
 ## v1.2.3
 * CMake changes to support package installation in conan.